requirement.txt
python-dotenv
requests
praw
Flask